+++ python/src/marple/utils.py
@@ -1,6 +1,6 @@
 import re
 from pathlib import Path
-from typing import Iterable, Literal
+from typing import Any, Iterable, Literal
 from urllib import parse, request
 
 import pandas as pd
@@ -20,7 +20,7 @@ from marple.db.constants import (
 )
 
 
-def validate_response(response: requests.Response, failure_message: str) -> dict:
+def validate_response(response: requests.Response, failure_message: str) -> Any:
     if response.status_code == 400:
         raise ValueError(f"{failure_message}: Bad request. {response.json().get('error', 'Unknown error')}")
     if response.status_code == 403: